Check  firstly if files exist before creating symbolic links

+++ cloudera/hdfs-rbf-parcel/src/main/scripts/create-tarball.sh
@@ -24,8 +24,12 @@ ln -sf "/etc/hadoop/conf" "conf"
 
 printlog "INFO: Creating symbolic links under lib/native"
 cd "${BUILD_DIRECTORY}/${PARCEL_BASENAME}/lib/hdfs_rbf/lib/native"
-ln -sf "libhdfs.so.0.0.0" "libhdfs.so"
-ln -sf "libhdfspp.so.0.1.0" "libhdfspp.so"
+if [[ -f "libhdfs.so.0.0.0" ]]; then
+  ln -sf "libhdfs.so.0.0.0" "libhdfs.so"
+fi
+if [[ -f "libhdfspp.so.0.1.0" ]]; then
+  ln -sf "libhdfspp.so.0.1.0" "libhdfspp.so"
+fi
 
 printlog "INFO: Changing to ${BUILD_DIRECTORY} and creating ${TAR_GZ_FILENAME} from ${PARCEL_BASENAME}"
 cd "${BUILD_DIRECTORY}"
